move EventSourcing.Serialization DI code into main EventSourcing.Serialization package
