+++ EventSourcing.Bus/Bus.RabbitMQ.IntegrationTests/Bus.RabbitMQ.IntegrationTests.csproj
@@ -23,6 +23,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\..\Core\EventSourcing\EventSourcing.csproj" />
       <ProjectReference Include="..\..\EventSourcing.Serialization\Extensions.DependencyInjection.Serialization.Json\Extensions.DependencyInjection.Serialization.Json.csproj" />
       <ProjectReference Include="..\..\Helpers\TestHelpers\TestHelpers.csproj" />
       <ProjectReference Include="..\Bus.RabbitMQ\Bus.RabbitMQ.csproj" />
