+++ EventSourcing.Serialization/Extensions.DependencyInjection.Serialization.Json/EventSourcingBuilderExtensions.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Text.Json;
 using EventSourcing.Abstractions.DependencyInjection;
+using EventSourcing.Serialization.Abstractions.DependencyInjection;
+using EventSourcing.Serialization.Extensions;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace EventSourcing.Extensions.DependencyInjection.Serialization.Json
