+++ EventSourcing.Serialization/Extensions.DependencyInjection.Serialization.Json/Extensions.DependencyInjection.Serialization.Json.csproj
@@ -27,9 +27,9 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Extensions.DependencyInjection.Serialization\Extensions.DependencyInjection.Serialization.csproj" />
         <ProjectReference Include="..\Serialization.Abstractions\Serialization.Abstractions.csproj" />
         <ProjectReference Include="..\Serialization.Json\Serialization.Json.csproj" />
+        <ProjectReference Include="..\Serialization\Serialization.csproj" />
     </ItemGroup>
 
 </Project>
