+++ EventSourcing.Serialization/Extensions.DependencyInjection.Serialization.NewtonsoftJson.UnitTests/Extensions.DependencyInjection.Serialization.NewtonsoftJson.UnitTests.csproj
@@ -21,6 +21,7 @@
     </ItemGroup>
 
     <ItemGroup>
+        <ProjectReference Include="..\..\Core\EventSourcing\EventSourcing.csproj" />
         <ProjectReference Include="..\..\Helpers\TestHelpers\TestHelpers.csproj" />
         <ProjectReference Include="..\Extensions.DependencyInjection.Serialization.NewtonsoftJson\Extensions.DependencyInjection.Serialization.NewtonsoftJson.csproj" />
     </ItemGroup>
