+++ EventSourcing.Serialization/Extensions.DependencyInjection.Serialization.NewtonsoftJson/Extensions.DependencyInjection.Serialization.NewtonsoftJson.csproj
@@ -27,9 +27,9 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Extensions.DependencyInjection.Serialization\Extensions.DependencyInjection.Serialization.csproj" />
         <ProjectReference Include="..\Serialization.Abstractions\Serialization.Abstractions.csproj" />
         <ProjectReference Include="..\Serialization.NewtonsoftJson\Serialization.NewtonsoftJson.csproj" />
+        <ProjectReference Include="..\Serialization\Serialization.csproj" />
     </ItemGroup>
 
 </Project>
