--- EventSourcing.Serialization/Extensions.DependencyInjection.Serialization/Extensions.DependencyInjection.Serialization.csproj
@@ -1,34 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <TreatWarningsAsErrors Condition=" '$(Configuration)' == 'Release' ">true</TreatWarningsAsErrors>
-        <GenerateDocumentationFile>true</GenerateDocumentationFile>
-        <RootNamespace>EventSourcing.Extensions.DependencyInjection.Serialization</RootNamespace>
-        <AssemblyName>EventSourcing.Extensions.DependencyInjection.Serialization</AssemblyName>
-    </PropertyGroup>
-
-    <!-- NuGet metadata -->
-    <PropertyGroup>
-        <Title>EventSourcing.Extensions.DependencyInjection.Serialization</Title>
-        <Description>This package contains DependencyInjection extensions for default implementation of EventSourcing.Serialization.Abstractions package.</Description>
-        <Authors>Dawid Szmidka</Authors>
-        <PackageId>EventSourcing.Extensions.DependencyInjection.Serialization</PackageId>
-        <Version>0.1.0-alpha</Version>
-        <PackageReleaseNotes>
-            - Initial version of EventSourcing.Extensions.DependencyInjection.Serialization project
-        </PackageReleaseNotes>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <PackageTags>EventSourcing;EventSourcing.Extensions;EventSourcing.Extensions.DependencyInjection;EventSourcing.Serialization</PackageTags>
-        <IncludeSymbols>true</IncludeSymbols>
-        <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-        <IncludeSource>true</IncludeSource>
-        <RepositoryUrl>https://github.com/TheName/EventSourcing</RepositoryUrl>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\..\Core\EventSourcing\EventSourcing.csproj" />
-        <ProjectReference Include="..\Serialization\Serialization.csproj" />
-    </ItemGroup>
-
-</Project>
