+++ EventSourcing.Serialization/Serialization.UnitTests/Serialization.UnitTests.csproj
@@ -7,7 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.1" />
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.0" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1" />
         <PackageReference Include="xunit" Version="2.4.1" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
@@ -21,8 +21,9 @@
     </ItemGroup>
 
     <ItemGroup>
+        <ProjectReference Include="..\..\Core\EventSourcing\EventSourcing.csproj" />
         <ProjectReference Include="..\..\Helpers\TestHelpers\TestHelpers.csproj" />
-        <ProjectReference Include="..\Extensions.DependencyInjection.Serialization\Extensions.DependencyInjection.Serialization.csproj" />
+        <ProjectReference Include="..\Serialization\Serialization.csproj" />
     </ItemGroup>
 
 </Project>
