+++ EventSourcing.Serialization/Serialization/Serialization.csproj
@@ -27,6 +27,10 @@
     </PropertyGroup>
     
     <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="5.0.0" />        
+    </ItemGroup>
+    
+    <ItemGroup>
       <ProjectReference Include="..\Serialization.Abstractions\Serialization.Abstractions.csproj" />
     </ItemGroup>
