+++ EventSourcing.sln
@@ -74,8 +74,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventSourcing.UnitTests.Ref
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serialization", "EventSourcing.Serialization\Serialization\Serialization.csproj", "{1450F357-8633-4F64-BB94-435D504108FB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extensions.DependencyInjection.Serialization", "EventSourcing.Serialization\Extensions.DependencyInjection.Serialization\Extensions.DependencyInjection.Serialization.csproj", "{C8F807A0-1770-4417-8410-61A2C7E89A58}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SqlServer", "SqlServer", "{CF15C20C-3512-4966-8BA8-20F6C98BB96E}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "PostgreSql", "PostgreSql", "{8B8FC55F-39CF-431D-AF1B-D67A24CB4071}"
@@ -89,8 +87,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.PostgreSql.IntegrationTests", "EventSourcing.Persistence\PostgreSql\Persistence.PostgreSql.IntegrationTests\Persistence.PostgreSql.IntegrationTests.csproj", "{246BB686-60C1-45E2-B435-719C55015E6F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.IntegrationTests.Base", "EventSourcing.Persistence\Persistence.IntegrationTests.Base\Persistence.IntegrationTests.Base.csproj", "{DC12019B-B5C2-4AB1-A0B1-89372FC05C48}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extensions.DependencyInjection.Serialization.UnitTests", "EventSourcing.Serialization\Extensions.DependencyInjection.Serialization.UnitTests\Extensions.DependencyInjection.Serialization.UnitTests.csproj", "{C92098C2-9DEB-466E-9AFB-D0F6700CDD85}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extensions.DependencyInjection.Serialization.Json.UnitTests", "EventSourcing.Serialization\Extensions.DependencyInjection.Serialization.Json.UnitTests\Extensions.DependencyInjection.Serialization.Json.UnitTests.csproj", "{F65145F4-5FE1-4A39-8710-9A1417BA5001}"
 EndProject
@@ -137,6 +133,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ForgettablePayloads.Persistence.PostgreSql.IntegrationTests", "EventSourcing.ForgettablePayloads\Persistence\PostgreSql\ForgettablePayloads.Persistence.PostgreSql.IntegrationTests\ForgettablePayloads.Persistence.PostgreSql.IntegrationTests.csproj", "{41B5FB23-BF1F-4F04-A278-B36478FC20B7}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ForgettablePayloads.Persistence.IntegrationTests.Base", "EventSourcing.ForgettablePayloads\Persistence\ForgettablePayloads.Persistence.IntegrationTests.Base\ForgettablePayloads.Persistence.IntegrationTests.Base.csproj", "{3BBA6BC9-4E9F-42F1-B318-02C2964F2459}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serialization.UnitTests", "EventSourcing.Serialization\Serialization.UnitTests\Serialization.UnitTests.csproj", "{9F02497F-C769-40A3-92D8-CC477428F5F0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -268,10 +266,6 @@ Global
 		{1450F357-8633-4F64-BB94-435D504108FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1450F357-8633-4F64-BB94-435D504108FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1450F357-8633-4F64-BB94-435D504108FB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C8F807A0-1770-4417-8410-61A2C7E89A58}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C8F807A0-1770-4417-8410-61A2C7E89A58}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C8F807A0-1770-4417-8410-61A2C7E89A58}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C8F807A0-1770-4417-8410-61A2C7E89A58}.Release|Any CPU.Build.0 = Release|Any CPU
 		{97B81C83-4D70-4B5D-98E9-D3E8590AF674}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{97B81C83-4D70-4B5D-98E9-D3E8590AF674}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{97B81C83-4D70-4B5D-98E9-D3E8590AF674}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -292,10 +286,6 @@ Global
 		{DC12019B-B5C2-4AB1-A0B1-89372FC05C48}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DC12019B-B5C2-4AB1-A0B1-89372FC05C48}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DC12019B-B5C2-4AB1-A0B1-89372FC05C48}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C92098C2-9DEB-466E-9AFB-D0F6700CDD85}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C92098C2-9DEB-466E-9AFB-D0F6700CDD85}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C92098C2-9DEB-466E-9AFB-D0F6700CDD85}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C92098C2-9DEB-466E-9AFB-D0F6700CDD85}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F65145F4-5FE1-4A39-8710-9A1417BA5001}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F65145F4-5FE1-4A39-8710-9A1417BA5001}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F65145F4-5FE1-4A39-8710-9A1417BA5001}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -372,6 +362,10 @@ Global
 		{3BBA6BC9-4E9F-42F1-B318-02C2964F2459}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3BBA6BC9-4E9F-42F1-B318-02C2964F2459}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3BBA6BC9-4E9F-42F1-B318-02C2964F2459}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9F02497F-C769-40A3-92D8-CC477428F5F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9F02497F-C769-40A3-92D8-CC477428F5F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9F02497F-C769-40A3-92D8-CC477428F5F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9F02497F-C769-40A3-92D8-CC477428F5F0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{FED5B0D5-163B-404E-9CD3-69C7B0A246B4} = {3829A2ED-C973-4C40-89B2-276296AC020A}
@@ -401,7 +395,6 @@ Global
 		{2CDB97C9-D7FD-4929-A99D-F3B7D8163BCF} = {554639EC-1D6A-41B9-8B71-07DE055A27EB}
 		{B23CEB74-9659-4E73-A264-8076BF2E69F1} = {051E2AFC-545F-4977-BA96-6BB80736FE2C}
 		{1450F357-8633-4F64-BB94-435D504108FB} = {29745774-1ACA-4563-89CD-8C145BAF9D29}
-		{C8F807A0-1770-4417-8410-61A2C7E89A58} = {29745774-1ACA-4563-89CD-8C145BAF9D29}
 		{CF15C20C-3512-4966-8BA8-20F6C98BB96E} = {8792BDC7-C64B-4270-A84C-579204AC79B5}
 		{1B426689-F3C7-460B-A8D1-539CDFE09630} = {CF15C20C-3512-4966-8BA8-20F6C98BB96E}
 		{6A449FAA-DB82-4988-B5BF-B4BAD8C4C9E2} = {CF15C20C-3512-4966-8BA8-20F6C98BB96E}
@@ -413,7 +406,6 @@ Global
 		{B9719B28-3265-4232-AB7E-04BB1CA55DA1} = {8B8FC55F-39CF-431D-AF1B-D67A24CB4071}
 		{246BB686-60C1-45E2-B435-719C55015E6F} = {8B8FC55F-39CF-431D-AF1B-D67A24CB4071}
 		{DC12019B-B5C2-4AB1-A0B1-89372FC05C48} = {8792BDC7-C64B-4270-A84C-579204AC79B5}
-		{C92098C2-9DEB-466E-9AFB-D0F6700CDD85} = {29745774-1ACA-4563-89CD-8C145BAF9D29}
 		{F65145F4-5FE1-4A39-8710-9A1417BA5001} = {29745774-1ACA-4563-89CD-8C145BAF9D29}
 		{6520A5D9-FD64-48E8-95B4-A047D543BBF1} = {29745774-1ACA-4563-89CD-8C145BAF9D29}
 		{74365D15-F56D-4410-BF63-D12D63694EDA} = {082DB854-6B33-49C4-A6FE-FDE3E5862FD8}
@@ -436,5 +428,6 @@ Global
 		{5FC89E6F-1E21-4345-BC29-6102C67D9D5C} = {C9014F64-CD59-46F5-9590-8CFAAC5B9E67}
 		{41B5FB23-BF1F-4F04-A278-B36478FC20B7} = {C9014F64-CD59-46F5-9590-8CFAAC5B9E67}
 		{3BBA6BC9-4E9F-42F1-B318-02C2964F2459} = {E9547AC6-1592-4CC0-9612-6A8629E23132}
+		{9F02497F-C769-40A3-92D8-CC477428F5F0} = {29745774-1ACA-4563-89CD-8C145BAF9D29}
 	EndGlobalSection
 EndGlobal
